Cast enumerator values of GL tokens to int.
C enumerator values must be convertible to signed int. The declared
values were unsigned instead.

+++ gdk/gdkgltokens.h
@@ -45,21 +45,21 @@ extern "C" {
 ** this file except in compliance with the License. You may obtain a copy
 ** of the License at Silicon Graphics, Inc., attn: Legal Services, 1600
 ** Amphitheatre Parkway, Mountain View, CA 94043-1351, or at:
-** 
+**
 ** http://oss.sgi.com/projects/FreeB
-** 
+**
 ** Note that, as provided in the License, the Software is distributed on an
 ** "AS IS" basis, with ALL EXPRESS AND IMPLIED WARRANTIES AND CONDITIONS
 ** DISCLAIMED, INCLUDING, WITHOUT LIMITATION, ANY IMPLIED WARRANTIES AND
 ** CONDITIONS OF MERCHANTABILITY, SATISFACTORY QUALITY, FITNESS FOR A
 ** PARTICULAR PURPOSE, AND NON-INFRINGEMENT.
-** 
+**
 ** Original Code. The Original Code is: OpenGL Sample Implementation,
 ** Version 1.2.1, released January 26, 2000, developed by Silicon Graphics,
 ** Inc. The Original Code is Copyright (c) 1991-2000 Silicon Graphics, Inc.
 ** Copyright in any portions created by third parties is as indicated
 ** elsewhere herein. All Rights Reserved.
-** 
+**
 ** Additional Notice Provisions: The application programming interfaces
 ** established by SGI in conjunction with the Original Code are The
 ** OpenGL(R) Graphics System: A Specification (Version 1.2.1), released
@@ -143,7 +143,7 @@ typedef enum
  */
 typedef enum
 {
-  GDK_GL_CONFIG_CAVEAT_DONT_CARE    = 0xFFFFFFFF, /* GDK_GL_DONT_CARE */
+  GDK_GL_CONFIG_CAVEAT_DONT_CARE    = (int)0xFFFFFFFF, /* GDK_GL_DONT_CARE */
   GDK_GL_CONFIG_CAVEAT_NONE         = 0x8000,     /* GDK_GL_NONE */
   GDK_GL_SLOW_CONFIG                = 0x8001,
   GDK_GL_NON_CONFORMANT_CONFIG      = 0x800D
@@ -155,7 +155,7 @@ typedef enum
  */
 typedef enum
 {
-  GDK_GL_VISUAL_TYPE_DONT_CARE      = 0xFFFFFFFF, /* GDK_GL_DONT_CARE */
+  GDK_GL_VISUAL_TYPE_DONT_CARE      = (int)0xFFFFFFFF, /* GDK_GL_DONT_CARE */
   GDK_GL_TRUE_COLOR                 = 0x8002,
   GDK_GL_DIRECT_COLOR               = 0x8003,
   GDK_GL_PSEUDO_COLOR               = 0x8004,
